+++ timezonefinder/helpers.py
@@ -20,7 +20,7 @@ def inside_polygon(x, y, coordinates):
         ng read directly from a binary file
     :return: true if the point (x,y) lies within the polygon
 
-    Some overflow considerations for the critical comparison of line segment slopes:
+    Some overflow considerations for the critical part of comparing the line segment slopes:
 
         (y2 - y) * (x2 - x1) <= delta_y_max * delta_x_max
         (y2 - y1) * (x2 - x) <= delta_y_max * delta_x_max
@@ -241,7 +241,8 @@ def distance_to_polygon_exact(lng_rad, lat_rad, nr_points, points, trans_points)
 
 
 def distance_to_polygon(lng_rad, lat_rad, nr_points, points):
-    # the maximum possible distance is half the perimeter of earth pi * 12743km = 40,054.xxx km
+    # the maximum possible distance is half the perimeter of earth
+    # equatorial circumference: 40075.017 km
     min_distance = 40100000
 
     for i in range(nr_points):